\name{as.matrix.by}
\alias{as.matrix.by}
\title{Coerces a by object into a matrix (only
tested on a 2d objects).}
\usage{
  \method{as.matrix}{by} (x, ...)
}
\arguments{
  \item{x}{is a by object to convert to a matrix}

  \item{\dots}{ignored}
}
\value{
  a matrix
}
\description{
  Coerces a by object into a matrix (only tested on a 2d
  objects).
}

