\name{last}
\alias{first}
\alias{last}
\title{Convenience functions to return the last/first element of a vector}
\usage{
  last(vec)
}
\arguments{
  \item{vec}{Vector of any type}
}
\value{
  Vector of length 1 of same type as vec
}
\description{
  Convenience functions to return the last/first element of
  a vector
}
\examples{
test <- seq(10)
first(test)
last(test)
}

