\name{subset.patteRns}
\alias{subset.patteRns}
\title{Determine which patterns a character vector utilizes}
\usage{
  subset.patteRns(x, ch, ...)
}
\arguments{
  \item{x}{An object of class patteRns}

  \item{ch}{The character vector to examine}

  \item{\dots}{Ignored}
}
\value{
  A patteRns object containing the matching patterns
}
\description{
  Determine which patterns a character vector utilizes
}

