\name{taRifx-package}
\alias{taRifx-package}
\alias{taRifx}
\docType{package}
\title{
Random collection of functions, particularly those for working with SpatialPolygons, SpatialPoints, etc. objects
}
\description{
Random collection of functions, particularly those for working with SpatialPolygons, SpatialPoints, etc. objects
}
\details{
\tabular{ll}{
Package: \tab taRifx\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-08\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ari B. Friedman 

Maintainer: <abfriedman@gmail.com>

}
\references{
None
}
\keyword{ package }
\examples{
bytable(runif(100),indices=list(rep(c('a','b'),50)))
}
