\name{xtable.CrossTable}
\alias{xtable.CrossTable}
\title{Add in methods to handle CrossTable objects in xtable}
\usage{
  \method{xtable}{CrossTable} (x, caption = NULL,
    label = NULL, align = NULL, digits = NULL,
    display = NULL, beta.names = NULL, ...)
}
\arguments{
  \item{x}{Model object}

  \item{caption}{Caption for table}

  \item{label}{See ?xtable}

  \item{align}{See ?xtable}

  \item{digits}{See ?xtable}

  \item{display}{See ?xtable}

  \item{beta.names}{See ?xtable}

  \item{\dots}{Arguments to pass to xtable}
}
\value{
  xtable object
}
\description{
  Add in methods to handle CrossTable objects in xtable
}
\seealso{
  \code{\link[xtable]{xtable}}
}

