\name{daysofweek}
\alias{daysofweek}
\title{Return a vector of the days of the week, in order}
\usage{
  daysofweek(start.day = "Monday")
}
\arguments{
  \item{start.day}{Day of the week to begin the week with
  (as a text item)}
}
\value{
  Character vector of length 7
}
\description{
  Return a vector of the days of the week, in order
}
\examples{
daysofweek("Sunday")
}

