\name{title.page.new}
\alias{title.page.new}
\title{Plot a title page containing the given text.  Good for breaking up sections
of plot PDFs.}
\usage{
  title.page.new(title.text = "")
}
\arguments{
  \item{title.text}{Text to plot on its own page}
}
\value{
  Plot
}
\description{
  Plot a title page containing the given text.  Good for
  breaking up sections of plot PDFs.
}
\examples{
title.page.new("Page break!")
}

