\name{xtable.lme}
\alias{xtable.lme}
\alias{xtable.summary.lme}
\title{Add in methods to handle LME objects in xtable}
\usage{
  xtable.lme(x, caption = NULL, label = NULL, align = NULL,
    digits = NULL, display = NULL, beta.names = NULL, ...)
}
\arguments{
  \item{x}{Model object}

  \item{caption}{Caption for table}

  \item{label}{See ?xtable}

  \item{align}{See ?xtable}

  \item{digits}{See ?xtable}

  \item{display}{See ?xtable}

  \item{beta.names}{See ?xtable}

  \item{\dots}{Arguments to pass to xtable}
}
\value{
  xtable object
}
\description{
  Add in methods to handle LME objects in xtable
}
\seealso{
  \code{\link[xtable]{xtable}}
}

